+++ game_style.css
@@ -48,6 +48,7 @@ body {
 }
 .level p {
     font-size: 16px;
+    margin-left: 0;
 }
 .rest {
     display: flex;
