+++ index.html
@@ -14,6 +14,7 @@
         />
         <link rel="stylesheet" href="styles/difficultyLevel.css" />
         <link rel="stylesheet" href="game_style.css"/>
+        <link rel="stylesheet" href="timer_style.css"/>
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Poppins" />
     </head>
     <body>
